Extract shared HashSubset helper from HashExcept + HashSlice
Create hash_subset.rs with HashSubsetMode enum and shared check_hash_subset
function. All 13 duplicated helper functions (block_param_names,
unwrap_negation, binary_other_side, include_collection, in_collection,
is_lvar_named, is_range_like, format_arg, etc.) are now in the shared
module. Each cop provides only the mode (Except vs Slice) that controls
the semantic matching logic.

HashExcept: 361 → 46 lines, HashSlice: 362 → 46 lines.
Includes CI cop-check dispatch for shared module changes.
