+++ scripts/dispatch_cops.py
@@ -91,6 +91,13 @@ VARIABLE_FORCE_CONSUMERS = {
     "Rails/SaveBang",
 }
 
+# Cops that consume the shared HashSubset helper. Changes to
+# src/cop/style/hash_subset.rs should trigger corpus checks for both.
+HASH_SUBSET_CONSUMERS = {
+    "Style/HashExcept",
+    "Style/HashSlice",
+}
+
 # Department PascalCase → snake_case directory name in src/cop/ and tests/fixtures/
 # Only needed for departments where pascal_to_snake() gives the wrong result
 DEPT_TO_SRC_DIR = {
@@ -1535,6 +1542,9 @@ def detect_cops(base: str, head: str) -> list[str]:
             if dept == "variable_force":
                 # Engine changes affect all VF consumer cops
                 cops.update(VARIABLE_FORCE_CONSUMERS)
+            elif name == "hash_subset":
+                # Shared helper changes affect both hash subset cops
+                cops.update(HASH_SUBSET_CONSUMERS)
             elif name not in {"mod", "node_type"}:
                 cops.add(f"{dept_snake_to_pascal(dept)}/{snake_to_pascal(name)}")
             continue
